Moved pages to follow Next.JS naming conventions
